Extract positive-Z nudge planning module
Move candidate detection and overused-edge confirmation into nudge_plan. It
sits above the cell class because confirming a contact builds a throwaway cell
and runs zero-height cleanup on it to see which contact edges actually survive.

Make build_positive_z_nudge_plan and its confirmation filter public, since
TouchTerrainEarthEngine drives both when generating an interlocking pair.

Import Callable from typing rather than collections.abc in the modules
extracted so far. Annotations are evaluated eagerly here, so the two spell
different objects and the moved functions were not carrying the annotations
they started with.

Emitted meshes are unchanged: four smoke configs stay byte identical.
